rename to mapCurrentIds
