+++ src/logic/mapCurrentIds.ts
